refactor(explore-task): 1.Refactor task handling methods for better readability and efficiency. 2.refactor tasklist to avoid thread conflicts
