+++ main.py
@@ -125,8 +125,8 @@ if __name__ == '__main__':
     # tt.solve("total_assault")
     # tt.solve("cafe_reward")
     # tt.solve("momo_talk")
-    # tt.solve("explore_normal_task")
-    tt.solve("explore_hard_task")
+    tt.solve("explore_normal_task")
+    # tt.solve("explore_hard_task")
     # tt.solve("normal_task")
     # tt.solve("hard_task")
     # tt.solve("arena")
