+++ module/ExploreTasks/__init__.py
@@ -1,0 +1,3 @@
+from module.ExploreTasks import explore_normal_task
+from module.ExploreTasks import explore_hard_task
+from module.ExploreTasks import TaskUtils
