+++ module/activities/PresidentHinasSummerVacation.py
@@ -2,7 +2,7 @@ from module.activities.activity_utils import get_stage_data
 import time
 from core import color, picture, image
 from module import main_story
-from module.ExploreTasks.explore_normal_task import common_gird_method
+from module.ExploreTasks.TaskUtils import execute_grid_task
 
 
 def implement(self):
@@ -156,7 +156,7 @@ def start_story(self, i):
     elif res == "reward_acquired":
         pass
     else:
-        common_gird_method(self, get_stage_data(self)["story" + str(i)])
+        execute_grid_task(self, get_stage_data(self)["story" + str(i)])
         main_story.auto_fight(self)
     return
 
@@ -223,7 +223,7 @@ def explore_challenge(self):
             elif res == "no-pass" or res == "pass":
                 need_fight = True
         if need_fight:
-            common_gird_method(self, current_task_stage_data)
+            execute_grid_task(self, current_task_stage_data)
             i += 1
         main_story.auto_fight(self)
         if self.config['manual_boss']:
