+++ module/activities/bunnyChaserOnTheShip.py
@@ -2,7 +2,7 @@ from module.activities.activity_utils import get_stage_data
 import time
 from core import color, picture
 from module import main_story
-from module.ExploreTasks.explore_normal_task import common_gird_method
+from module.ExploreTasks.TaskUtils import execute_grid_task
 
 
 def implement(self):
@@ -199,7 +199,7 @@ def explore_mission(self):
                 need_fight = True
         if need_fight:
             self.logger.info("Start mission " + str(task_number) + " fight")
-            common_gird_method(self, current_task_stage_data)
+            execute_grid_task(self, current_task_stage_data)
             main_story.auto_fight(self)
             if self.config['manual_boss']:
                 self.click(1235, 41)
@@ -249,7 +249,7 @@ def explore_challenge(self):
                     elif res == "no-pass" or res == "pass":
                         need_fight = True
                 if need_fight:
-                    common_gird_method(self, current_task_stage_data)
+                    execute_grid_task(self, current_task_stage_data)
                     i += 1
                 main_story.auto_fight(self)
                 if self.config['manual_boss']:
